Separate LoginForm from LoginFormContainer
