+++ src/LoginForm.jsx
@@ -1,0 +1,47 @@
+import React from 'react';
+
+export default function LoginForm({ fields, onChange }) {
+  const { email, password } = fields;
+
+  function handleChange(event) {
+    const { target: { name, value } } = event;
+
+    onChange({ name, value });
+  }
+
+  return (
+    <>
+      <div>
+        <label htmlFor="login-email">
+          E-mail
+        </label>
+        <input
+          type="email"
+          id="login-email"
+          name="email"
+          value={email}
+          onChange={handleChange}
+        />
+      </div>
+      <div>
+        <label htmlFor="login-password">
+          Password
+        </label>
+        <input
+          type="password"
+          id="login-password"
+          name="password"
+          value={password}
+          onChange={handleChange}
+        />
+      </div>
+      <div>
+        <button
+          type="button"
+        >
+          Log In
+        </button>
+      </div>
+    </>
+  );
+}
