+++ src/LoginFormContainer.jsx
@@ -2,6 +2,8 @@ import React from 'react';
 
 import { useDispatch, useSelector } from 'react-redux';
 
+import LoginForm from './LoginForm';
+
 import { changeLoginField } from './actions';
 
 import { get } from './utils';
@@ -11,45 +13,14 @@ export default function LoginFormContainer() {
 
   const { email, password } = useSelector(get('loginFields'));
 
-  function handleChange(event) {
-    const { target: { name, value } } = event;
-
+  function handleChange({ name, value }) {
     dispatch(changeLoginField({ name, value }));
   }
 
   return (
-    <>
-      <div>
-        <label htmlFor="login-email">
-          E-mail
-        </label>
-        <input
-          type="email"
-          id="login-email"
-          name="email"
-          value={email}
-          onChange={handleChange}
-        />
-      </div>
-      <div>
-        <label htmlFor="login-password">
-          Password
-        </label>
-        <input
-          type="password"
-          id="login-password"
-          name="password"
-          value={password}
-          onChange={handleChange}
-        />
-      </div>
-      <div>
-        <button
-          type="button"
-        >
-          Log In
-        </button>
-      </div>
-    </>
+    <LoginForm
+      fields={{ email, password }}
+      onChange={handleChange}
+    />
   );
 }
